Moved tests pkg under util
